File: add the file management through a block design
[Definition]
A block is a page-sized region from the log file. It is represented as a
timestamp through the date of its first and last log.

[Design]
The main idea behind this pattern is to represent the file as a readonly
sorted array of timestamp. Therefore, the search of a date is done in
a log(N) complexity where N is the number of blocks.
This is also interesting due to the fact that it is not really greedy in
memory: a thread map only the page it is looking into.

[Optimizations]
* The block search is not parallelize. There should be no lock needed.
* The search of the dates delimiting a block is done when the block is
  created. This is currently the most time consuming action. It could be
  really interesting to vectorize it.

Signed-off-by: Alexandre Bernard <alexandre.bernard@lse.epita.fr>