file: add unordered file support
The block design implemented in 5303d3ee1fb59555776e5a92d66daacabf07084a
cannot be applied to unordered files. This commit add the UnorderedLogFile
class.

[Design]
The file is parsed on load in order to build a mapping between a timing
and the associated queries. The parsing is done in a O(N) complexity where
N is the number of bytes in the file.
The search is done in O(N) complexity where N is the number of requests.

[Optimizations]
Currently the bottleneck is the use of cache unfriendly structure
(unordered_map) and string comparaisons.
* Radix tree with the date (represented as a strict group of 48 bits)
  as index.
  2018-05-29 00:47:12 would be: 0010 0000 0001 1000 (0x2018)
                                0000 0101 0010 1001 (0x0529)
                                0000 0000 0100 0111 (0x0047)
                                0001 0010           (0x12)

Signed-off-by: Alexandre Bernard <alexandre.bernard@lse.epita.fr>